Fix services to run on the API

+++ services/document-processor/services/processor/models.py
@@ -7,7 +7,7 @@ from decimal import Decimal
 class Transaction:
     date: date
     description: str
-    amount: Decimal
+    amount: float
     raw_text: str = ""
     confidence: float = 0.0
 
@@ -17,7 +17,7 @@ class Transaction:
         return cls(
             date=current_date or date.today(),
             description=line_data.get('description', ''),
-            amount=Decimal(str(line_data.get('amount', 0))),
+            amount=float(str(line_data.get('amount', 0))),
             raw_text=' '.join(w['text'] for w in line_data.get('words', []))
         )
